backend/requirements.txt
Flask>=3.0,<4.0
psycopg[binary]>=3.2,<4.0
PyJWT[crypto]>=2.10,<3.0
python-dotenv>=1.0,<2.0
